pyproject.toml
[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"
[project]
name = "RedFox-Futebol-Backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel", email = "danielmoggi2019@gmail.com"},
]
dependencies = [
    "django>=4.2.7",
    "netifaces>=0.11.0",
    "setuptools>=69.0.2",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.1",
    "djangorestframework-simplejwt>=5.3.0",
    "python-magic>=0.4.27",
    "Pillow>=10.1.0",
    "drf-spectacular>=0.26.5",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
    "whitenoise[brotli]>=6.6.0",
    "psycopg2-binary>=2.9.9",
    "django-cloudinary-storage>=0.3.0",
    "django-filter>=23.4",
    "django-extensions>=3.2.3",
    "pydotplus>=2.0.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"




[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.12.0",
    "black>=23.11.0",
]
